pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terracost"
version = "0.1.0"
description = "A tool for estimating Terraform infrastructure costs with uncertainty analysis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Daniël van Zyl"},
    {name = "Shailyn Ramsamy Moodley"},
    {name = "Tevlen Naidoo"}
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "requests>=2.25.0",
    "python-hcl2>=7.0.0",
    "pydantic>=2.0.0",
    "boto3>=1.26.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
terracost = "terracost.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["terracost*"]
